InterpreterWPF: Add ExportToString converter to make some properties more human readable for objectproperty. Print ByteProperty, NoneProperty(Grey). Added getting ArrayType. Implemented ToggleHexboxWidth and set the proper sizes. Hooked up method for re-ordering arrayproperty (not yet implemented).
PropertyCollection: If requesting properties with inclusion of None, properties are forced to reload. This it to keep existing behavior where inclusion of none is default to false (InterpreterWPF will use true).

PackageEditorWPF: Export changed status now uses standardized scan of all exports instead of custom ones per header, data. This will use more CPU but will standardize the implementation.

ExportEntry: If header is the same as the previous, the setter exits, ignoring all side effects and not actually changing the data pointer. GetProperties() will now only cache the properties if forceReload is false, which is default. This will keep backwards compatibility with other code so that NoneProperties are not cached.

ME3UnrealObjectInfo: If export.idxClass <= 0, generating class info is ignored as this would cause an exception.